+++ app/controllers/application_controller.rb
@@ -49,6 +49,30 @@ class ApplicationController < ActionController::API
     render json: { error: exception.message }, status: :unprocessable_entity
   end
 
+  private def slot_paging_params
+    p = params.permit(:filter, :moment, :limit, :after, :before).symbolize_keys
+
+    # are there any pagination params?
+    return {} unless p.any?
+
+    # set default limit if not provided
+    p[:limit] = PAGINATION_DEFAULT_LIMIT if p[:limit].nil?
+    # set maximum for limit to 100 if higher
+    p[:limit] = PAGINATION_MAX_LIMIT if p[:limit].to_i > PAGINATION_MAX_LIMIT
+
+    # ignore status & moment if a cursor is submitted
+    if p[:before].present? || p[:after].present?
+      p[:filter] = nil
+      p[:moment] = nil
+    else
+      # set default status and moment if not provided
+      p[:filter] = PAGINATION_DEFAULT_FILTER if p[:filter].nil?
+      p[:moment] = Time.zone.now.to_s if p[:moment].nil?
+    end
+    p
+  end
+
+  # for Pundit
   class UserContext
     attr_reader :current_user, :requested_user
 
@@ -67,6 +91,7 @@ class ApplicationController < ActionController::API
     end
   end
 
+  # I18n
   def set_locale
     I18n.locale = (current_user[:lang] || I18n.default_locale)
   end