+++ app/models/base_slot.rb
@@ -1,4 +1,5 @@
 class BaseSlot < ActiveRecord::Base
+  include SlotFollow
   # this class is not intended to be used directly
   # but rather as an uniform interface for the specific slot representations
   # it shares postgres inheritance semantics at the db level with its subtypes
@@ -29,6 +30,7 @@ class BaseSlot < ActiveRecord::Base
   after_initialize :set_slot_type, if: :new_record?
 
   scope :active, -> { where deleted_at: nil }
+  # there are additonal scopes defined as class method (upcoming, past)
   # there is also a default scope defined as class method
 
   has_many :media_items, -> { where deleted_at: nil }, as: :mediable
@@ -260,7 +262,8 @@ class BaseSlot < ActiveRecord::Base
   ## class methods ##
 
   def self.default_scope
-    where(slot_type: SLOT_TYPES[to_s.to_sym]) unless self == BaseSlot
+    # apply default scope only to the 'leaves'
+    where(slot_type: SLOT_TYPES[to_s.to_sym]) if subclasses.empty?
   end
 
   def self.get(slot_id)
@@ -272,11 +275,15 @@ class BaseSlot < ActiveRecord::Base
     slot_ids.collect { |id| get(id) }
   end
 
-  # TODO: add spec, check performance
+  # TODO: add spec
   def self.upcoming
     # BaseSlot.includes(:meta_slot).
       # where('meta_slots.start_date > ?', Time.zone.now).references(:meta_slot)
     joins(:meta_slot).where('meta_slots.start_date > ?', Time.zone.now)
+  end
+
+  def self.past
+    joins(:meta_slot).where('meta_slots.start_date < ?', Time.zone.now)
   end
 
   def self.next
@@ -359,8 +366,8 @@ class BaseSlot < ActiveRecord::Base
     # the following is not really neccessary, might be removed at some point
     # but for now it gives some useful info about the system
     else
-      if slot.start_date.to_s != cursor[:startdate] ||
-         slot.end_date.to_s != cursor[:enddate]
+      if slot.start_date.strftime('%Y-%m-%d %H:%M:%S.%N') != cursor[:startdate] ||
+         slot.end_date.strftime('%Y-%m-%d %H:%M:%S.%N') != cursor[:enddate]
         opts = {}
         opts[:parameters] = { cursor_id: cursor[:id],
                               cursor_startdate: cursor[:startdate],