+++ app/models/comment.rb
@@ -1,4 +1,4 @@
-class Comment < ActiveRecord::Base
+class Comment < SlotActivity
   after_commit AuditLog
 
   belongs_to :slot, class_name: BaseSlot, inverse_of: :comments
@@ -9,5 +9,27 @@ class Comment < ActiveRecord::Base
 
   def delete
     update(deleted_at: Time.zone.now)
+  end
+
+  ## Activity Methods ##
+
+  private
+
+  def activity_slot
+    slot
+  end
+
+  def activity_user
+    user
+  end
+
+  def activity_verb
+    'comment'
+  end
+
+  # The message is used as a notification message
+  # for the users activity feed
+  def activity_message
+    "#{I18n.t('activity_create_comment', title: activity_slot.meta_slot.title)}"
   end
 end