+++ app/models/group_follow.rb
@@ -1,0 +1,8 @@
+module GroupFollow
+
+  include Follow
+
+  def feed_type
+    'Group'
+  end
+end