+++ app/models/like.rb
@@ -1,4 +1,4 @@
-class Like < ActiveRecord::Base
+class Like < SlotActivity
   after_commit AuditLog
 
   belongs_to :user
@@ -8,5 +8,27 @@ class Like < ActiveRecord::Base
 
   def delete
     update(deleted_at: Time.zone.now)
+  end
+
+  ## Activity Methods ##
+
+  private
+
+  def activity_slot
+    slot
+  end
+
+  def activity_user
+    user
+  end
+
+  def activity_verb
+    'like'
+  end
+
+  # The message is used as a notification message
+  # for the users activity feed
+  def activity_message
+    "#{I18n.t('activity_create_like', title: activity_slot.meta_slot.title)}"
   end
 end